Fixed semgrep warning

+++ lib/server/resources/abstract/host.go
@@ -88,7 +88,7 @@ func almostEqual(a, b float32) bool {
 }
 
 func almostEqual64(a, b float64) bool {
-	return math.Abs(float64(a-b)) <= 1e-6
+	return math.Abs(a-b) <= 1e-6
 }
 
 func (hsr HostSizingRequirements) Equals(in HostSizingRequirements) bool {
